refactor: file structure in weather models
