+++ src/models/weather-condition.ts
@@ -1,0 +1,6 @@
+export type WeatherCondition = {
+  id: number;
+  main: string;
+  description: string;
+  icon: string;
+};
