+++ src/models/weather-summary.ts
@@ -1,3 +1,11 @@
+import {
+  CurrentWeatherDescription,
+  DailyWeatherDescription,
+  HourlyWeatherDescription,
+  MinutelyWeatherDescription,
+  WeatherAlert,
+} from './weather-data';
+
 export type WeatherSummary = {
   lat: number;
   lon: number;
@@ -8,95 +16,4 @@ export type WeatherSummary = {
   hourly: HourlyWeatherDescription[];
   daily: DailyWeatherDescription[];
   alerts: WeatherAlert[];
-};
-
-export type WeatherCondition = {
-  id: number;
-  main: string;
-  description: string;
-  icon: string;
-};
-
-export type CurrentWeatherDescription = {
-  dt: number;
-  sunrise: number;
-  sunset: number;
-  temp: number;
-  feels_like: number;
-  pressure: number;
-  humidity: number;
-  dew_point: number;
-  uvi: number;
-  clouds: number;
-  visibility: number;
-  wind_speed: number;
-  wind_deg: number;
-  wind_gust: number;
-  weather: WeatherCondition[];
-};
-
-export type MinutelyWeatherDescription = {
-  dt: number;
-  precipitation: number;
-};
-
-export type HourlyWeatherDescription = {
-  dt: number;
-  temp: number;
-  feels_like: number;
-  pressure: number;
-  humidity: number;
-  dew_point: number;
-  uvi: number;
-  clouds: number;
-  visibility: number;
-  wind_speed: number;
-  wind_deg: number;
-  wind_gust: number;
-  weather: WeatherCondition[];
-  pop: number;
-};
-
-export type DailyWeatherDescription = {
-  dt: number;
-  sunrise: number;
-  sunset: number;
-  moonrise: number;
-  moonset: number;
-  moon_phase: number;
-  summary: string;
-  temp: {
-    day: number;
-    min: number;
-    max: number;
-    night: number;
-    eve: number;
-    morn: number;
-  };
-  feels_like: {
-    day: number;
-    night: number;
-    eve: number;
-    morn: number;
-  };
-  pressure: number;
-  humidity: number;
-  dew_point: number;
-  wind_speed: number;
-  wind_deg: number;
-  wind_gust: number;
-  weather: WeatherCondition[];
-  clouds: number;
-  pop: number;
-  rain: number;
-  uvi: number;
-};
-
-export type WeatherAlert = {
-  sender_name: string;
-  event: string;
-  start: number;
-  end: number;
-  description: string;
-  tags: string[];
 };
